# [web:canvaskit] switch to temporary SkPaint objects